+++ Runtime/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Unity.XR.Management.Tests")]
+[assembly: InternalsVisibleTo("Unity.XR.Management.EditorTests")]