+++ Runtime/XRManagementAnalytics.cs
@@ -1,34 +1,30 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Diagnostics.CodeAnalysis;
-
 #if UNITY_EDITOR
 using UnityEditor;
 #endif
-
 #if UNITY_ANALYTICS && ENABLE_CLOUD_SERVICES_ANALYTICS
 using UnityEngine.Analytics;
-#endif //UNITY_ANALYTICS && ENABLE_CLOUD_SERVICES_ANALYTICS
+#endif
 
 [assembly:InternalsVisibleTo("Unity.XR.Management.Editor")]
 namespace UnityEngine.XR.Management
 {
-    internal static class XRManagementAnalytics
+    static class XRManagementAnalytics
     {
-        private const int kMaxEventsPerHour = 1000;
-        private const int kMaxNumberOfElements = 1000;
-        private const string kVendorKey = "unity.xrmanagement";
-        private const string kEventBuild = "xrmanagment_build";
+        const int k_MaxEventsPerHour = 1000;
+        const int k_MaxNumberOfElements = 1000;
+        const string k_VendorKey = "unity.xrmanagement";
+        const string k_EventBuild = "xrmanagment_build";
 
 #if ENABLE_CLOUD_SERVICES_ANALYTICS && UNITY_ANALYTICS
-        private static bool s_Initialized = false;
+        static bool s_Initialized = false;
 #endif //ENABLE_CLOUD_SERVICES_ANALYTICS && UNITY_ANALYTICS
 
         [Serializable]
-        private struct BuildEvent
+        struct BuildEvent
 #if UNITY_2023_2_OR_NEWER && ENABLE_CLOUD_SERVICES_ANALYTICS && UNITY_ANALYTICS
             : IAnalytic.IData
 #endif //UNITY_2023_2_OR_NEWER && ENABLE_CLOUD_SERVICES_ANALYTICS && UNITY_ANALYTICS
@@ -40,26 +36,30 @@ namespace UnityEngine.XR.Management
         }
 
 #if UNITY_2023_2_OR_NEWER && ENABLE_CLOUD_SERVICES_ANALYTICS && UNITY_ANALYTICS
-        [AnalyticInfo(eventName: kEventBuild, vendorKey: kVendorKey, maxEventsPerHour: kMaxEventsPerHour, maxNumberOfElements: kMaxNumberOfElements)]
-        private class XrInitializeAnalytic : IAnalytic
+        [AnalyticInfo(
+            eventName: k_EventBuild,
+            vendorKey: k_VendorKey,
+            maxEventsPerHour: k_MaxEventsPerHour,
+            maxNumberOfElements: k_MaxNumberOfElements)]
+        class XrInitializeAnalytic : IAnalytic
         {
-            private BuildEvent? data = null;
+            BuildEvent? m_Data;
 
             public XrInitializeAnalytic(BuildEvent data)
             {
-                this.data = data;
+                m_Data = data;
             }
 
             public bool TryGatherData(out IAnalytic.IData data, [NotNullWhen(false)] out Exception error)
             {
                 error = null;
-                data = this.data;
+                data = m_Data;
                 return data != null;
             }
         }
 #endif //UNITY_2023_2_OR_NEWER && ENABLE_CLOUD_SERVICES_ANALYTICS && UNITY_ANALYTICS
 
-        private static bool Initialize()
+        static bool Initialize()
         {
 #if ENABLE_TEST_SUPPORT || !ENABLE_CLOUD_SERVICES_ANALYTICS || !UNITY_ANALYTICS
             return false;
@@ -71,27 +71,25 @@ namespace UnityEngine.XR.Management
             if (!EditorAnalytics.enabled)
                 return false;
 
-            if(AnalyticsResult.Ok != EditorAnalytics.RegisterEventWithLimit(kEventBuild, kMaxEventsPerHour, kMaxNumberOfElements, kVendorKey))
+            if(AnalyticsResult.Ok != EditorAnalytics.RegisterEventWithLimit(k_EventBuild, k_MaxEventsPerHour, k_MaxNumberOfElements, k_VendorKey))
                 return false;
             s_Initialized = true;
 #endif //UNITY_EDITOR
             return s_Initialized;
 #endif //ENABLE_TEST_SUPPORT || !ENABLE_CLOUD_SERVICES_ANALYTICS || !UNITY_ANALYTICS
-
         }
 
 #if UNITY_EDITOR
-        public static void SendBuildEvent(GUID guid, BuildTarget buildTarget, BuildTargetGroup buildTargetGroup, IEnumerable<XRLoader> loaders)
+        public static void SendBuildEvent(
+            GUID guid, BuildTarget buildTarget, BuildTargetGroup buildTargetGroup, IEnumerable<XRLoader> loaders)
         {
-
 #if UNITY_ANALYTICS && ENABLE_CLOUD_SERVICES_ANALYTICS
-
             if (!s_Initialized && !Initialize())
             {
                 return;
             }
 
-            List<string> loaderTypeNames = new List<string>();
+            var loaderTypeNames = new List<string>();
             foreach (var loader in loaders)
             {
                 loaderTypeNames.Add(loader.GetType().Name);
@@ -107,10 +105,9 @@ namespace UnityEngine.XR.Management
 #if UNITY_2023_2_OR_NEWER
             EditorAnalytics.SendAnalytic(new XrInitializeAnalytic(data));
 #else
-            EditorAnalytics.SendEventWithLimit(kEventBuild, data);
+            EditorAnalytics.SendEventWithLimit(k_EventBuild, data);
 #endif //UNITY_2023_2_OR_NEWER
 #endif //UNITY_ANALYTICS && ENABLE_CLOUD_SERVICES_ANALYTICS
-
         }
 #endif //UNITY_EDITOR
     }